feat: Enhance AI Analyzer with local analysis and JSON repair features
- Added methods for local analysis of transcripts, extracting keywords and action items.
- Implemented JSON repair functionality to handle malformed responses from Ollama.
- Updated the AIAnalyzer class to prioritize technical terms and keywords without duplication.
- Improved error handling and logging for JSON decoding issues.
- Refactored the summarization process to streamline interactions with the Ollama API.

fix: Update SpeakerDiarizer to use modern token API for Hugging Face

- Changed the token parameter in the Pipeline initialization to align with the latest Hugging Face API.

feat: Enhance SpeechRecognizer with configurable thresholds and chunk processing

- Introduced parameters for no-speech and logprob thresholds.
- Implemented chunk processing for long audio files to improve transcription accuracy.
- Updated transcribe method to handle new parameters and improve performance.

chore: Update Docker Compose configuration for improved service management

- Renamed services for clarity and updated PostgreSQL version.
- Adjusted environment variables for database connection.
- Configured GPU support for the AI service.

chore: Update requirements.txt to include new dependencies

- Added huggingface-hub and updated other dependencies for compatibility.

feat: Add recent meetings retrieval endpoint in MeetingController

- Implemented a new endpoint to fetch the most recent meetings.
- Updated MeetingRepository to support fetching the latest meetings.

fix: Configure application properties for proper encoding in MeetingService

- Set UTF-8 encoding for Tomcat and servlet to ensure correct character handling.

feat: Add CORS support to ProcessingController and MeetingController

- Enabled CORS for frontend compatibility during development.

+++ ai-service/app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, DateTime, Float, ForeignKey, JSON
+from sqlalchemy import Column, Integer, BigInteger, String, Text, DateTime, Float, ForeignKey, JSON
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from app.database import Base
@@ -8,7 +8,7 @@ class Transcript(Base):
     __tablename__ = "transcripts"
     
     id = Column(Integer, primary_key=True, index=True)
-    meeting_id = Column(Integer, nullable=False, index=True)
+    meeting_id = Column(BigInteger, nullable=False, index=True)
     speaker = Column(String(50))
     start_time = Column(Float)
     end_time = Column(Float)
@@ -23,7 +23,7 @@ class Analysis(Base):
     __tablename__ = "analysis"
     
     id = Column(Integer, primary_key=True, index=True)
-    meeting_id = Column(Integer, nullable=False, unique=True, index=True)
+    meeting_id = Column(BigInteger, nullable=False, unique=True, index=True)
     summary = Column(Text)
     keywords = Column(JSON)  # List of keywords
     technical_terms = Column(JSON)  # List of technical terms
